+++ src/UserInterfaceConstants.hpp
@@ -22,6 +22,7 @@ const size_t NumColourSchemes = 3;
 
 static const char* _ecv_array const axisNames[] = { "X", "Y", "Z", "U", "V", "W" };
 constexpr size_t MaxTotalAxes = 15;		// This needs to be kept in sync with the maximum in RRF for any build configuration
+constexpr size_t MaxHeatersPerTool = 8;
 static const char* _ecv_array const babystepAmounts[] = { "0.01", "0.02", "0.05", "0.1" };
 static const float _ecv_array babystepAmountsF[] = { 0.01, 0.02, 0.05, 0.1 };